Added %failed_categorization_type_count to count how many of each of the various kinds of failed macro categorizations we get.  These are not really well integrated with the categorizations yet.
Print the above hash as a histogram to the STATS stream

Call check_macro_body() on the failed categorizations, and let it update
the above hash.

Fix check_macro_body() -- it had an endless loop due to an extraneous my
declaration.  Have it return a result (like before--unused) and the
failure_type, which is then used in updating the hash.