Many editing changes, including the following.
Use epsfig, which is documented in the _LaTeX Companion_, instead of epsf
(which I've never heard of before and can't find documentation for).  In
order to do this, I had to convert to LaTeX2e, which involved no more than
changing "documentstyle" to "documentclass".

Title "of the use of the C Preprocessor" => "of C Preprocessor Use"

I substantially expanded the figure captions; every figure should stand on
its own without requiring the reader to hunt through the paper for
explanations.  (See _Scientific American_ for good caption examples.)

Please don't use "etc.", because it makes us look like we can't figure out
a descriptive generalization for the concept we are trying to explain.

Replaced most uses of \verb with {\tt ...}.  This works in moving arguments
and often produces nicer spacing.  \verb is still most expedient when the
argument includes underscores or braces.  Use \# to get a literal
octothorpe, in \tt or in roman font.