Made these suggested fixes from David's email (see diffs for actual changes made)
*	"to an invocation of the function intended" => "to an invocation of
foo, the function intended"
*	After "virtually all C programs." include a sentence like: "In
addition, the framework we describe could be instantiated for these other
uses of the preprocessor in a fairly direct way."
*	No, you can keep the numbered sections, if you want (either way is
OK).
*	"testing C code extensions is justified" => "testing C code
extensions is acceptable"
*	"A substantial strength of" => "However, a substantial strength of"
*	Footnote #8 (that the tool is available) should surely be moved
in-text, I'd say as a parenthetical.
*	"is automatically generate-able and" => "could be generated
automatically and"


Still have to do this one:
*	I think it's a good idea to add the parameters to the hooks in the
appendix, if it can be down in a somewhat simple way.