Fixed date.
Minor typos corrected.

Quote pronunciation ``pee-see-pee-cubed.''

Improve the layout of the arguments in Appendix A;  give (..) as args
in heading line when they are described in the paragraph (done when
they are too long to fit, or especially repetitive w/ the other
hooks mentioned).

This is ready to go by me...